LUA: Use 32-bits LUAC file format on all platforms
Lua scripts in The Witcher were compiled using a 32-bit compiler.

+++ lua/llimits.h
@@ -128,7 +128,7 @@ typedef LUA_UACNUMBER l_uacNumber;
 ** type for virtual-machine instructions
 ** must be an unsigned with (at least) 4 bytes (see details in lopcodes.h)
 */
-typedef unsigned long Instruction;
+typedef uint32_t Instruction;
 
 
 /* maximum depth for calls (unsigned short) */
